requirements.txt
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.1
Wikipedia-API==0.6.0
